requirements.txt
torch~=2.1.1
numpy~=1.26.2
torchvision~=0.16.1
wandb~=0.16.0
matplotlib~=3.8.2
tqdm~=4.66.1
pandas~=2.1.3
Pillow~=10.1.0